+++ tests/tests_cycle_decomposition/test_cases.py
@@ -36,6 +36,10 @@ TEST_CYCLE_DECOMPOSITION = [
         CycleDecomposition(Cycle(1), Cycle(2, 4, 7, 6), Cycle(3, 5)),
     ),
 ]
+TEST_INVERSE = [
+    (CycleDecomposition(Cycle(1, 2, 3)), CycleDecomposition(Cycle(3, 2, 1))),
+    (CycleDecomposition(Cycle(1, 2), Cycle(3, 4)), CycleDecomposition(Cycle(2, 1), Cycle(4, 3))),
+]
 TEST_IS_DERANGEMENT = [
     (CycleDecomposition(Cycle(1)), False),
     (CycleDecomposition(Cycle(1), Cycle(2, 4, 7, 6), Cycle(3, 5)), False),